+++ composer.json
@@ -30,7 +30,8 @@
     "require-dev": {
         "phpunit/phpunit": "^7.3",
         "symfony/browser-kit": "^4.1",
-        "symfony/debug": "^4.1"
+        "symfony/debug": "^4.1",
+        "symfony/phpunit-bridge": "^4.1"
     },
     "replace": {
         "symfony/polyfill-ctype": "*",
