+++ config/bundles.php
@@ -1,6 +1,7 @@
 <?php
 
 return [
+    Libero\HttpClientBundle\HttpClientBundle::class => ['all' => true],
     Libero\ContentPageBundle\ContentPageBundle::class => ['all' => true],
     Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
 ];
