+++ symfony.lock
@@ -155,6 +155,15 @@
     "symfony/http-kernel": {
         "version": "v4.1.3"
     },
+    "symfony/phpunit-bridge": {
+        "version": "4.1",
+        "recipe": {
+            "repo": "github.com/symfony/recipes",
+            "branch": "master",
+            "version": "4.1",
+            "ref": "1a47f76dcd9d5d4da1fab54fb15450bae3704c5e"
+        }
+    },
     "symfony/polyfill-mbstring": {
         "version": "v1.9.0"
     },
